PlateformeCollecteDonnees/src/client/requirements.txt
mysql-connector-python
pyserial
opencv-python
depthai
numpy